wip: moved the life time validation function back to the class
